chore: rework gradle plugins to support desktop apps.
