+++ build-logic/plugins/src/main/kotlin/conventions/ApplicationPlugin.kt
@@ -1,0 +1,16 @@
+package conventions
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class ApplicationPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target) {
+            with(pluginManager) {
+                apply("org.jetbrains.kotlin.multiplatform")
+                apply("org.jetbrains.compose")
+                apply("org.jetbrains.kotlin.plugin.compose")
+            }
+        }
+    }
+}
